Move CV information to Hugo data dir to that Hugo's livereload will work
By default Hugo will watch for changes when running `hugo server`.
The configuration (e.g. `config.toml`) is not watched and as a result
changing CV content in the config file will not get livereloaded.
This commit fixes that.

As data seems to not be read from theme directories, this does however
introduce the need for people testing this theme to copy the
exampleSite/data/cv.toml to hugoroot/data for test-running the theme.